+++ tests/TestCase/Shell/Task/AppUsesTaskTest.php
@@ -25,7 +25,7 @@ class AppUsesTaskTest extends TestCase {
 	 *
 	 * @return void
 	 */
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$io = $this->getMockBuilder(ConsoleIo::class)->getMock();
