+++ tests/TestCase/Shell/Task/FixtureLoadingTaskTest.php
@@ -25,7 +25,7 @@ class FixtureLoadingTaskTest extends TestCase {
 	 *
 	 * @return void
 	 */
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$io = $this->getMockBuilder(ConsoleIo::class)->getMock();
