+++ tests/TestCase/Shell/Task/I18nTaskTest.php
@@ -23,7 +23,7 @@ class I18nTaskTest extends TestCase {
 	 *
 	 * @return void
 	 */
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$io = $this->getMockBuilder(ConsoleIo::class)->getMock();
