+++ tests/TestCase/Snippets/MethodSignatures/ControllerSnippetsTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Cake\Upgrade\Test\TestCase\Utility;
+namespace Cake\Upgrade\Test\TestCase\Snippets\MethodSignatures;
 
 use Cake\TestSuite\TestCase;
 use Cake\Upgrade\Shell\Task\ChangeTrait;
-use Cake\Upgrade\Utility\ComponentSnippets;
+use Cake\Upgrade\Snippets\MethodSignatures\ControllerSnippets;
 
-class ComponentSnippetsTest extends TestCase {
+class ControllerSnippetsTest extends TestCase {
 
 	use ChangeTrait;
 
@@ -14,7 +14,7 @@ class ComponentSnippetsTest extends TestCase {
 	 * @return void
 	 */
 	public function testReplace(): void {
-		$replacements = (new ComponentSnippets())->events();
+		$replacements = (new ControllerSnippets())->events();
 
 		$string = 'public function beforeFilter(Event $event)';
 		$result = $this->exec($string, $replacements);
@@ -28,6 +28,17 @@ class ComponentSnippetsTest extends TestCase {
 		$string = 'public function beforeFilter(EventInterface $event): void';
 		$result = $this->exec($string, $replacements);
 		$this->assertSame($string, $result);
+	}
+
+	/**
+	 * @return void
+	 */
+	public function testPlugins(): void {
+		$replacements = (new ControllerSnippets())->plugins();
+		$string = '$this->loadComponent(\'Search.Prg\');';
+		$result = $this->exec($string, $replacements);
+		$expected = '$this->loadComponent(\'Search.Search\');';
+		$this->assertSame($expected, $result);
 	}
 
 	/**
